+++ packages/cached_query_flutter/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.3.0+2
+
+- Update Cached Query to the latest release.
+
+## 0.3.0+1
+
+ - **DOCS**: add base projects and infinite query with bloc.
+
 ## 0.3.0
 
 > Note: This release has breaking changes.